feat(app): fetch posts from `/api/posts`
