--- app/src/data/users.ts
@@ -1,28 +1,0 @@
-export interface UserData {
-  name: string;
-  avatar: string;
-}
-
-export const amiraHassan: UserData = {
-  name: "Amira Hasan",
-  avatar:
-    "https://images.unsplash.com/photo-1544005313-94ddf0286df2?w=100&h=100&fit=crop&crop=face",
-};
-
-export const budiSantoso: UserData = {
-  name: "Budi Santoso",
-  avatar:
-    "https://images.unsplash.com/photo-1506794778202-cad84cf45f1d?w=100&h=100&fit=crop&crop=face",
-};
-
-export const sitiNurhaliza: UserData = {
-  name: "Siti Nurhaliza",
-  avatar:
-    "https://images.unsplash.com/photo-1508214751196-bcfd4ca60f91?w=100&h=100&fit=crop&crop=face",
-};
-
-export const ranggaPratama: UserData = {
-  name: "Rangga Pratama",
-  avatar:
-    "https://images.unsplash.com/photo-1547425260-76bcadfb4f2c?w=100&h=100&fit=crop&crop=face",
-};
